+++ src/main/java/com/example/taste/domain/party/controller/PartyController.java
@@ -20,10 +20,10 @@ import org.springframework.web.bind.annotation.RestController;
 import com.example.taste.common.annotation.ValidEnum;
 import com.example.taste.common.response.CommonResponse;
 import com.example.taste.config.security.CustomUserDetails;
-import com.example.taste.domain.party.dto.reponse.PartyDetailResponseDto;
-import com.example.taste.domain.party.dto.reponse.PartyResponseDto;
 import com.example.taste.domain.party.dto.request.PartyCreateRequestDto;
 import com.example.taste.domain.party.dto.request.PartyDetailUpdateRequestDto;
+import com.example.taste.domain.party.dto.response.PartyDetailResponseDto;
+import com.example.taste.domain.party.dto.response.PartyResponseDto;
 import com.example.taste.domain.party.enums.PartyFilter;
 import com.example.taste.domain.party.service.PartyInvitationService;
 import com.example.taste.domain.party.service.PartyService;
