+++ src/main/java/com/example/taste/domain/party/controller/PartyInvitationController.java
@@ -19,10 +19,10 @@ import org.springframework.web.bind.annotation.RestController;
 import com.example.taste.common.exception.CustomException;
 import com.example.taste.common.response.CommonResponse;
 import com.example.taste.config.security.CustomUserDetails;
-import com.example.taste.domain.party.dto.reponse.PartyInvitationResponseDto;
-import com.example.taste.domain.party.dto.reponse.UserInvitationResponseDto;
 import com.example.taste.domain.party.dto.request.InvitationActionRequestDto;
 import com.example.taste.domain.party.dto.request.PartyInvitationRequestDto;
+import com.example.taste.domain.party.dto.response.PartyInvitationResponseDto;
+import com.example.taste.domain.party.dto.response.UserInvitationResponseDto;
 import com.example.taste.domain.party.enums.InvitationStatus;
 import com.example.taste.domain.party.enums.InvitationType;
 import com.example.taste.domain.party.service.PartyInvitationService;
