+++ src/main/java/com/example/taste/domain/party/dto/response/UserInvitationResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.taste.domain.party.dto.reponse;
+package com.example.taste.domain.party.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
